Rename instrument classes
