+++ python_il_sts/utils/file_saving.py
@@ -14,7 +14,7 @@ from datetime import datetime
 
 # Importing STS process and instrument classes
 from python_il_sts.measurements.sts_process import StsProcess
-from python_il_sts.instruments.tsl_instrument_class import TslInstrument
+from python_il_sts.instruments.tsl_instrument import TslInstrument
 from python_il_sts.utils.error_handling_class import sts_process_error_strings, STSProcessError
 
 now = datetime.now()        # Get the current date and time
